feat: multiple agents can now be created
