+++ lib/agenix.dart
@@ -14,5 +14,4 @@ export 'src/memory/data/data_store.dart';
 // Tools
 export 'src/tools/tool.dart';
 export 'src/tools/param_spec.dart';
-export 'src/tools/tool_registry.dart';
 export 'src/tools/tool_response.dart';
