+++ lib/src/agent/agent.dart
@@ -1,9 +1,9 @@
 import 'dart:convert';
-
 import 'package:agenix/agenix.dart';
 import 'package:agenix/src/static/_pkg_constants.dart';
 import 'package:agenix/src/tools/_parser.dart';
 import 'package:agenix/src/tools/_tool_runner.dart';
+import 'package:agenix/src/tools/tool_registry.dart';
 import 'package:flutter/services.dart';
 
 part '_memory_manager.dart';
@@ -12,58 +12,54 @@ part '_prompt_builder.dart';
 /// Agent is the main class that represents the AI agent.
 /// Define the agent with all background knowledge and tools.
 class Agent {
-  static final Agent _instance = Agent._internal();
+  final _MemoryManager _memoryManager;
+  final _PromptBuilder _promptBuilder;
+  final PromptParser _promptParser = PromptParser();
 
-  /// Singleton
-  factory Agent() => _instance;
-  Agent._internal();
+  /// This is where you define the tools that the agent can use, registry allows you to register and unregister tools.
+  final ToolRegistry toolRegistry;
+  final ToolRunner _toolRunner = ToolRunner();
 
-  bool _isInitialized = false;
+  /// The LLM that powers the agent, you can use a pre-built model like Gemini or if you have a custom implementation running on the server, you can use that.
+  final LLM llm;
 
-  late final _MemoryManager _memoryManager;
-  late final _PromptBuilder _promptBuilder;
+  Agent._internal({
+    required this.llm,
+    required _MemoryManager memoryManager,
+    required _PromptBuilder promptBuilder,
+    required this.toolRegistry,
+  }) : _memoryManager = memoryManager,
+       _promptBuilder = promptBuilder;
 
-  /// The large language model that the agent uses can be any LLM, right now gemini is supported out of the box
-  late final LLM llm;
-  late final PromptParser _promptParser;
-  late final ToolRunner _toolRunner;
-
-  /// Initialize the agent with a data store and LLM, and load the system data.
-  /// In the system data you can define the personality of the agent, the name of the agent, and anything else that comes to mind.
-  Future<void> init({
+  /// Async factory constructor to create an instance with loaded system data.
+  static Future<Agent> create({
     required DataStore dataStore,
     required LLM llm,
-    String? pathToSystemData,
+    String pathToSystemData = 'assets/system_data.json',
   }) async {
-    if (_isInitialized) {
-      throw Exception('Agent is already initialized');
-    }
+    final systemData = await _loadSystemData(pathToSystemData);
+    final registry = ToolRegistry();
+    return Agent._internal(
+      llm: llm,
+      memoryManager: _MemoryManager(dataStore: dataStore),
+      promptBuilder: _PromptBuilder(
+        systemPrompt: systemData,
+        registry: registry,
+      ),
+      toolRegistry: registry,
+    );
+  }
+
+  static Future<Map<String, dynamic>> _loadSystemData(String path) async {
     try {
-      _isInitialized = true;
-
-      _memoryManager = _MemoryManager(dataStore: dataStore);
-      this.llm = llm;
-
-      final String jsonString = await rootBundle.loadString(
-        pathToSystemData ?? 'assets/system_data.json',
-      );
-      final raw = json.decode(jsonString);
-      _promptBuilder = _PromptBuilder(systemPrompt: raw);
-
-      _promptParser = PromptParser();
-      _toolRunner = ToolRunner();
+      final raw = await rootBundle.loadString(path);
+      return json.decode(raw);
     } catch (e) {
-      _isInitialized = false;
-      throw Exception('Failed to initialize Agent: $e');
+      throw Exception('Failed to load system data from $path: $e');
     }
   }
 
   /// Generate a response to the user message.
-  /// The response is generated by the LLM based on the context of the conversation.
-  /// The context is built from the conversation history and the system data.
-  /// The response is then passed to the tool runner to execute any tools that are needed.
-  /// The response is then saved to the memory manager.
-  /// The response is returned as an AgentMessage object.
   Future<AgentMessage> generateResponse({
     required String convoId,
     required AgentMessage userMessage,
@@ -87,7 +83,9 @@ class Agent {
         prompt: prompt,
         rawData: userMessage.imageData,
       );
+
       final parsed = _promptParser.parse(rawLLMResponse);
+
       if (parsed.toolNames.isEmpty) {
         final response = parsed.fallbackResponse ?? kLLMResponseOnFailure;
         final botResponse = AgentMessage(
@@ -99,10 +97,8 @@ class Agent {
         return botResponse;
       }
 
-      final toolResponses = await _toolRunner.runTools(parsed);
-
-      // TODO: Add support for tool chaining, right now we are just joining the messages.
-      String response = toolResponses.map((r) => r.message).join('\n');
+      final toolResponses = await _toolRunner.runTools(parsed, toolRegistry);
+      final response = toolResponses.map((r) => r.message).join('\n');
 
       final botMessage = AgentMessage(
         content: response.isEmpty ? kLLMResponseOnFailure : response,
@@ -112,9 +108,6 @@ class Agent {
       _memoryManager.saveMessage(convoId, botMessage);
       return botMessage;
     } catch (e) {
-      if (!_isInitialized) {
-        throw Exception('Agent is not initialized, please call init() first.');
-      }
       return AgentMessage(
         content: kLLMResponseOnFailure,
         isFromAgent: true,
@@ -123,36 +116,34 @@ class Agent {
     }
   }
 
-  /// Use the method to get all the messages in a conversation.
-  /// The messages are returned as a list of AgentMessage objects.
+  /// Get messages for a specific conversation from the datastore.
   Future<List<AgentMessage>> getMessages({
     required String conversationId,
     Object? metaData,
-  }) async {
-    return await _memoryManager.dataStore.getMessages(
+  }) {
+    return _memoryManager.dataStore.getMessages(
       conversationId,
       metaData: metaData,
     );
   }
 
-  /// Use the method to get all conversations for a user.
-  /// This allows developers to manage conversations and retrieve them as needed.
+  /// Get all conversations from the datastore.
   Future<List<Conversation>> getAllConversations({
     required String conversationId,
     Object? metaData,
-  }) async {
-    return await _memoryManager.dataStore.getConversations(
+  }) {
+    return _memoryManager.dataStore.getConversations(
       conversationId,
       metaData: metaData,
     );
   }
 
-  /// Use the method to delete a conversation.
+  /// Delete a conversation from the datastore.
   Future<void> deleteConversation({
     required String conversationId,
     Object? metaData,
-  }) async {
-    await _memoryManager.dataStore.deleteConversation(
+  }) {
+    return _memoryManager.dataStore.deleteConversation(
       conversationId,
       metaData: metaData,
     );
